feat(messages of user): got a v-for but next: delete some items from firestore to ease debugging.
